refactor: move shared packges to pkg folder
