+++ cmd/migration/main.go
@@ -93,7 +93,7 @@ func main() {
 		"sslmode=disable",
 	)
 
-	m, err := migrate.New(fmt.Sprintf("file:///%s/internal/db/migration", migrateFolder), dsn)
+	m, err := migrate.New(fmt.Sprintf("file:///%s/pkg/db/migration", migrateFolder), dsn)
 
 	if err != nil {
 		panic(err)
