+++ pkg/datamodel/jsonschema.go
@@ -10,7 +10,7 @@ import (
 	"google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/types/known/structpb"
 
-	"github.com/instill-ai/connector-backend/internal/logger"
+	"github.com/instill-ai/connector-backend/pkg/logger"
 )
 
 // SrcConnDefJSONSchema represents the SourceConnectorDefinition JSON Schema
