+++ pkg/worker/write.go
